stox-agent/requirements.txt

# Core dependencies
requests>=2.31.0
pathlib2>=2.3.7; python_version < '3.4'

# gRPC and protobuf
grpcio==1.74.0
protobuf>=4.25.3,<5.0.0

# Data processing
typing-extensions>=4.7.1

# Langchain dependencies
langchain>=0.1.0,<0.3.0
langchain-google-genai>=1.0.0,<3.0.0
langchain-core>=0.1.0,<0.3.0
google-generativeai>=0.3.0,<0.9.0
google-genai>=0.3.0,<0.9.0

# Web framework
flask>=2.3.0
python-dotenv>=1.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0